+++ aider/io.py
@@ -231,10 +231,17 @@ class InputOutput:
         self.user_input(inp)
         return inp
 
-    def add_to_file_history(self, inp):
+    def add_to_input_history(self, inp):
         if not self.input_history_file:
             return
         FileHistory(self.input_history_file).append_string(inp)
+
+    def get_input_history(self):
+        if not self.input_history_file:
+            return []
+
+        fh = FileHistory(self.input_history_file)
+        return fh.load_history_strings()
 
     def user_input(self, inp, log_only=True):
         if not log_only: