+++ HISTORY.md
@@ -1,5 +1,13 @@
 # Release history
 
+### main branch
+
+- [Only git commit dirty files that GPT tries to edit](https://github.com/paul-gauthier/aider/issues/200#issuecomment-1682750798)
+- Send chat history as prompt/context for Whisper voice transcription
+- Added `--voice-language` switch to constrain `/voice` to transcribe to a specific language
+- Late-bind importing `sounddevice`, as it was slowing down aider startup
+- Improved --foo/--no-foo switch handling for command line and yml config settings
+
 ### v0.12.0
 
 - [Voice-to-code](https://aider.chat/docs/voice.html) support, which allows you to code with your voice.